auth check and clearing user post relation

+++ controllers/feed.js
@@ -1,5 +1,6 @@
 const {validationResult}=require('express-validator/check');
 const Post=require('../models/post');
+const User=require('../models/user');
 const fs=require('fs');
 const path=require('path');
 
@@ -58,21 +59,29 @@ exports.createPost=(req,res,next)=>{
     //parsing data from the incoming request
     const title=req.body.title;
     const content=req.body.content; 
+    let creator;
     const post=new Post({
         title:title,
         content:content,
         //imageUrl:'image/rwby.jgp',
         imageUrl:imageUrl,
-        creator:{name:'Auz'}
+        creator:req.userId
             
     });
     post.save().then(result=>{
-        console.log(result);
+        return User.findById(req.userId);
+    }).then(user=>{
+        creator =user;
+        user.posts.push(post);
+        return user.save();
+    }).then(result=>{
         res.status(201).json({
             message:'Post created',
-            post:result
+            post:post,
+            creator:{_id:creator._id,name:creator.name}
         });
-    }).catch(err=>{
+    })
+    .catch(err=>{
         if(!err.statusCode){
             err.statusCode=500;
         }
@@ -126,6 +135,12 @@ exports.updatePost=(req,res,next)=>{
             error.statusCode=404;
             throw error;
         }
+        if(post.creator.toString()!=req.userId){
+            const error=new Error("Not authorized");
+            error.statusCode=403;
+            throw error;
+        }
+
         if(!imageUrl!==post.imageUrl){
             clearImage(post.imageUrl);
         }
@@ -152,13 +167,23 @@ exports.deletePost=(req,res,next)=>{
             error.statusCode=404;
             throw error;
         }
+        if(post.creator.toString()!=req.userId){
+            const error=new Error("Not authorized");
+            error.statusCode=403;
+            throw error;
+        }
         //check logged in user
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
     }).then(result=>{
-        console.log(result);
+        return User.findById(req.userId);
+    })  .then(user=>{
+        user.posts.pull(postId);
+        return user.save();
+    }).then(result=>{
+
         res.status(200).json({message:'Deleted'});
-    })  
+    })
     .catch(err=>{
         if(!err.statusCode){
             err.statusCode=500;
